+++ axon-event-commons/src/main/java/com/mycompany/axoneventcommons/order/OrderSubmittedEvent.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,6 +13,6 @@ public class OrderSubmittedEvent implements OrderEvent {
 
     private String orderId;
     private String status;
-    private Float total;
+    private BigDecimal total;
 
 }
