+++ food-ordering-service/src/main/java/com/mycompany/foodorderingservice/order/rest/dto/OrderDto.java
@@ -3,6 +3,7 @@ package com.mycompany.foodorderingservice.order.rest.dto;
 import com.mycompany.foodorderingservice.order.model.OrderStatus;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Data
@@ -15,7 +16,7 @@ public class OrderDto {
     private String restaurantId;
     private String restaurantName;
     private OrderStatus status;
-    private Float total;
+    private BigDecimal total;
     private Set<OrderItem> items;
 
     @Data
@@ -23,7 +24,7 @@ public class OrderDto {
         private String id;
         private String dishId;
         private String dishName;
-        private Float dishPrice;
+        private BigDecimal dishPrice;
         private Short quantity;
     }
